Moving GpuResources.h and creating a GPU math module.
