+++ src/pspg/field/FFT.h
@@ -16,7 +16,7 @@
 #include <cufft.h>
 #include <cuda.h>
 #include <cuda_runtime.h>
-#include <pspg/GpuResources.h>
+#include <pspg/math/GpuResources.h>
 
 namespace Pscf {
 namespace Pspg {
