+++ src/pspg/math/math.mod
@@ -1,0 +1,12 @@
+
+namespace Pscf{
+
+   /**
+   * \defgroup Pscf_Math_Module GPU Mathematics
+   *
+   * Mathematical utility classes specific to the CUDA implementation.
+   *
+   * \ingroup Pscf_Pspg_Module
+   */
+
+}
