+++ src/pspg/tests/cuda/CudaMemTest.h
@@ -6,7 +6,7 @@
 
 #include <pspg/field/RDField.h>
 #include <util/math/Constants.h>
-#include <pspg/GpuResources.h>
+#include <pspg/math/GpuResources.h>
 
 #include <fstream>
 #include <iomanip>
