+++ src/pspg/tests/solvers/PropagatorTest.h
@@ -12,7 +12,7 @@
 #include <pscf/math/IntVec.h>
 #include <util/math/Constants.h>
 
-#include <pspg/GpuResources.h>
+#include <pspg/math/GpuResources.h>
 #include <fstream>
